[chore

+++ app/API/version2/meetups/views.py
@@ -44,28 +44,24 @@ class DeleteMeetups(Resource):
     Class to handle deleting of meetups
     DELETE /api/v2/meetups/<int:meetup_Id> -> deletes a meetup
     """
-   #@jwt_required
+    @jwt_required
     def delete(self, meetup_id):
-        #if get_jwt_identity() != "admin12@gmail.com":
-            #return{
-                #"status": 403,
-                #"message": "You are not authorized to create a meetup"
-            #}, 403
-        #"""Route to delete a meetup"""
+        if get_jwt_identity() != "admin12@gmail.com":
+            return{
+                "status": 403,
+                "message": "You are not authorized to create a meetup"
+            }, 403
+        """Route to delete a meetup"""
         return Meetups.delete_meetups(self, meetup_id)
+
 
 class GetOneMeetupWithQuestions(Resource):
     """
     Class to handle fetching a specific meetup record
     GET /api/v2/meetups/<int:meetup_id>/questions -> Fetches a specific meetup 
     """
-    #@jwt_required
+    @jwt_required
     def get(self, meetup_id):
-        #if get_jwt_identity() != "admin12@gmail.com":
-            #return{
-                #"status": 403,
-                #"message": "You are not authorized to create a meetup"
-            #}, 403
         """Route to fetch a specific meetup"""
         return Meetups().get_one_meetup_with_questions(meetup_id)
 
@@ -74,7 +70,7 @@ class GetOneMeetup(Resource):
     Class to handle fetching a specific meetup record
     GET /api/v2/meetups/<int:meetup_id> -> Fetches a specific meetup 
     """
+    @jwt_required
     def get(self, meetup_id):
         """Route to fetch a specific meetup"""
         return Meetups().get_one_meetup(meetup_id)
-
